Split SA1117 unit tests into partical classes by target type.
