+++ StyleCop.Analyzers/StyleCop.Analyzers.Test/ReadabilityRules/SA1117/SA1117UnitTests.Attributes.cs
@@ -1,0 +1,66 @@
+﻿// Copyright (c) Tunnel Vision Laboratories, LLC. All Rights Reserved.
+// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
+
+namespace StyleCop.Analyzers.Test.ReadabilityRules
+{
+    using System.Threading;
+    using System.Threading.Tasks;
+    using StyleCop.Analyzers.ReadabilityRules;
+    using TestHelper;
+    using Xunit;
+
+    /// <summary>
+    /// Unit tests for attributes part of <see cref="SA1117ParametersMustBeOnSameLineOrSeparateLines"/>.
+    /// </summary>
+    public partial class SA1117UnitTests
+    {
+        [Fact]
+        public async Task TestValidAttributeAsync()
+        {
+            var testCode = @"
+[System.AttributeUsage(System.AttributeTargets.Class)]
+public class MyAttribute : System.Attribute
+{
+    public MyAttribute(int a, int b, int c)
+    {
+    }
+}
+
+[MyAttribute(1, 2, 3)]
+class Foo
+{
+}
+
+// This is a regression test for https://github.com/DotNetAnalyzers/StyleCopAnalyzers/issues/1211
+[System.Obsolete]
+class ObsoleteType
+{
+}";
+
+            await this.VerifyCSharpDiagnosticAsync(testCode, EmptyDiagnosticResults, CancellationToken.None).ConfigureAwait(false);
+        }
+
+        [Fact]
+        public async Task TestInvalidAttributeAsync()
+        {
+            var testCode = @"
+[System.AttributeUsage(System.AttributeTargets.Class)]
+public class MyAttribute : System.Attribute
+{
+    public MyAttribute(int a, int b, int c)
+    {
+    }
+}
+
+[MyAttribute(
+    1,
+    2, 3)]
+class Foo
+{
+}";
+
+            DiagnosticResult expected = this.CSharpDiagnostic().WithLocation(12, 8);
+            await this.VerifyCSharpDiagnosticAsync(testCode, expected, CancellationToken.None).ConfigureAwait(false);
+        }
+    }
+}
