+++ StyleCop.Analyzers/StyleCop.Analyzers.Test/ReadabilityRules/SA1117/SA1117UnitTests.ConstructorInitializers.cs
@@ -1,0 +1,81 @@
+﻿// Copyright (c) Tunnel Vision Laboratories, LLC. All Rights Reserved.
+// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
+
+namespace StyleCop.Analyzers.Test.ReadabilityRules
+{
+    using System.Collections.Generic;
+    using System.Threading;
+    using System.Threading.Tasks;
+    using StyleCop.Analyzers.ReadabilityRules;
+    using TestHelper;
+    using Xunit;
+
+    /// <summary>
+    /// Unit tests for constructor initializers part of <see cref="SA1117ParametersMustBeOnSameLineOrSeparateLines"/>.
+    /// </summary>
+    public partial class SA1117UnitTests
+    {
+        public static IEnumerable<object[]> GetTestConstructorInitializers(string delimiter)
+        {
+            yield return new object[] { $"this(42, 43, {delimiter} \"hello\")" };
+            yield return new object[] { $"base(42, 43, {delimiter} \"hello\")" };
+        }
+
+        [Theory]
+        [MemberData(nameof(GetTestConstructorInitializers), "")]
+        public async Task TestValidConstructorInitializerAsync(string initializer)
+        {
+            var testCode = $@"
+class Base
+{{
+    public Base(int a, int b, string s)
+    {{
+    }}
+}}
+
+class Derived : Base
+{{
+    public Derived()
+        : {initializer}
+    {{
+    }}
+
+    public Derived(int i, int j, string z)
+        : base(i, j, z)
+    {{
+    }}
+}}";
+
+            await this.VerifyCSharpDiagnosticAsync(testCode, EmptyDiagnosticResults, CancellationToken.None).ConfigureAwait(false);
+        }
+
+        [Theory]
+        [MemberData(nameof(GetTestConstructorInitializers), "\r\n")]
+        public async Task TestInvalidConstructorInitializerAsync(string initializer)
+        {
+            var testCode = $@"
+class Base
+{{
+    public Base(int a, int b, string s)
+    {{
+    }}
+}}
+
+class Derived : Base
+{{
+    public Derived()
+        : {initializer}
+    {{
+    }}
+
+    public Derived(int i, int j, string z)
+        : base(i, j, z)
+    {{
+    }}
+}}";
+
+            DiagnosticResult expected = this.CSharpDiagnostic().WithLocation(13, 2);
+            await this.VerifyCSharpDiagnosticAsync(testCode, expected, CancellationToken.None).ConfigureAwait(false);
+        }
+    }
+}
