+++ StyleCop.Analyzers/StyleCop.Analyzers.Test/ReadabilityRules/SA1117/SA1117UnitTests.Declarations.cs
@@ -1,0 +1,69 @@
+﻿// Copyright (c) Tunnel Vision Laboratories, LLC. All Rights Reserved.
+// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
+
+namespace StyleCop.Analyzers.Test.ReadabilityRules
+{
+    using System.Collections.Generic;
+    using System.Threading;
+    using System.Threading.Tasks;
+    using StyleCop.Analyzers.ReadabilityRules;
+    using TestHelper;
+    using Xunit;
+
+    /// <summary>
+    /// Unit tests for declarations part of <see cref="SA1117ParametersMustBeOnSameLineOrSeparateLines"/>.
+    /// </summary>
+    public partial class SA1117UnitTests
+    {
+        public static IEnumerable<object[]> GetTestDeclarations(string delimiter)
+        {
+            yield return new object[] { $"public Foo(int a, int b,{delimiter} string s) {{ }}" };
+            yield return new object[] { $"public object Bar(int a, int b,{delimiter} string s) => null;" };
+            yield return new object[] { $"public object this[int a, int b,{delimiter} string s] => null;" };
+            yield return new object[] { $"public delegate void Bar(int a, int b,{delimiter} string s);" };
+        }
+
+        public static IEnumerable<object[]> ValidTestDeclarations()
+        {
+            yield return new object[]
+            {
+                $@"public Foo(
+    int a, int b, string s) {{ }}"
+            };
+            yield return new object[]
+            {
+                $@"public Foo(
+    int a,
+    int b,
+    string s) {{ }}"
+            };
+        }
+
+        [Theory]
+        [MemberData(nameof(GetTestDeclarations), "")]
+        [MemberData(nameof(ValidTestDeclarations))]
+        public async Task TestValidDeclarationAsync(string declaration)
+        {
+            var testCode = $@"
+class Foo
+{{
+    {declaration}
+}}";
+            await this.VerifyCSharpDiagnosticAsync(testCode, EmptyDiagnosticResults, CancellationToken.None).ConfigureAwait(false);
+        }
+
+        [Theory]
+        [MemberData(nameof(GetTestDeclarations), "\r\n")]
+        public async Task TestInvalidDeclarationAsync(string declaration)
+        {
+            var testCode = $@"
+class Foo
+{{
+    {declaration}
+}}";
+
+            DiagnosticResult expected = this.CSharpDiagnostic().WithLocation(5, 2);
+            await this.VerifyCSharpDiagnosticAsync(testCode, expected, CancellationToken.None).ConfigureAwait(false);
+        }
+    }
+}
